rename PrivacyTransactionResponse to PrivacySendResponse
