+++ ethereum/api/src/test/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/privacy/methods/eea/EeaSendRawTransactionTest.java
@@ -36,7 +36,7 @@ import org.hyperledger.besu.ethereum.eth.transactions.TransactionPool;
 import org.hyperledger.besu.ethereum.mainnet.TransactionValidator.TransactionInvalidReason;
 import org.hyperledger.besu.ethereum.mainnet.ValidationResult;
 import org.hyperledger.besu.ethereum.privacy.PrivacyController;
-import org.hyperledger.besu.ethereum.privacy.PrivacyTransactionResponse;
+import org.hyperledger.besu.ethereum.privacy.PrivacySendResponse;
 import org.hyperledger.besu.ethereum.privacy.PrivateTransaction;
 import org.hyperledger.besu.util.bytes.BytesValue;
 
@@ -199,7 +199,7 @@ public class EeaSendRawTransactionTest {
   @Test
   public void validTransactionIsSentToTransactionPool() {
     when(privateTxHandler.sendTransaction(any(PrivateTransaction.class)))
-        .thenReturn(new PrivacyTransactionResponse(MOCK_ORION_KEY, MOCK_PRIVACY_GROUP));
+        .thenReturn(new PrivacySendResponse(MOCK_ORION_KEY, MOCK_PRIVACY_GROUP));
     when(privateTxHandler.validatePrivateTransaction(
             any(PrivateTransaction.class), any(String.class)))
         .thenReturn(ValidationResult.valid());
@@ -232,7 +232,7 @@ public class EeaSendRawTransactionTest {
   @Test
   public void validTransactionPrivacyGroupIsSentToTransactionPool() {
     when(privateTxHandler.sendTransaction(any(PrivateTransaction.class)))
-        .thenReturn(new PrivacyTransactionResponse(MOCK_ORION_KEY, MOCK_PRIVACY_GROUP));
+        .thenReturn(new PrivacySendResponse(MOCK_ORION_KEY, MOCK_PRIVACY_GROUP));
     when(privateTxHandler.validatePrivateTransaction(
             any(PrivateTransaction.class), any(String.class)))
         .thenReturn(ValidationResult.valid());
@@ -350,7 +350,7 @@ public class EeaSendRawTransactionTest {
       throws Exception {
 
     when(privateTxHandler.sendTransaction(any(PrivateTransaction.class)))
-        .thenReturn(new PrivacyTransactionResponse(MOCK_ORION_KEY, MOCK_PRIVACY_GROUP));
+        .thenReturn(new PrivacySendResponse(MOCK_ORION_KEY, MOCK_PRIVACY_GROUP));
     when(privateTxHandler.validatePrivateTransaction(
             any(PrivateTransaction.class), any(String.class)))
         .thenReturn(ValidationResult.valid());
