+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/privacy/PrivacySendResponse.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package org.hyperledger.besu.ethereum.privacy;
+
+public class PrivacySendResponse {
+  private final String enclaveKey;
+  private final String privacyGroup;
+
+  public PrivacySendResponse(final String enclaveKey, final String privacyGroup) {
+    this.enclaveKey = enclaveKey;
+    this.privacyGroup = privacyGroup;
+  }
+
+  public String getEnclaveKey() {
+    return enclaveKey;
+  }
+
+  public String getPrivacyGroup() {
+    return privacyGroup;
+  }
+}
