+++ ethereum/core/src/test/java/org/hyperledger/besu/ethereum/privacy/PrivacyControllerTest.java
@@ -162,16 +162,15 @@ public class PrivacyControllerTest {
 
     final PrivateTransaction transaction = buildLegacyPrivateTransaction(1);
 
-    final PrivacyTransactionResponse privacyTransactionResponse =
-        privacyController.sendTransaction(transaction);
+    final PrivacySendResponse privacySendResponse = privacyController.sendTransaction(transaction);
 
     final ValidationResult<TransactionInvalidReason> validationResult =
         privacyController.validatePrivateTransaction(
-            transaction, privacyTransactionResponse.getPrivacyGroup());
+            transaction, privacySendResponse.getPrivacyGroup());
 
     final Transaction markerTransaction =
         privacyController.createPrivacyMarkerTransaction(
-            privacyTransactionResponse.getEnclaveKey(), transaction);
+            privacySendResponse.getEnclaveKey(), transaction);
 
     assertThat(validationResult).isEqualTo(ValidationResult.valid());
     assertThat(markerTransaction.contractAddress()).isEqualTo(PUBLIC_TRANSACTION.contractAddress());
@@ -186,8 +185,7 @@ public class PrivacyControllerTest {
 
     final PrivateTransaction transaction = buildBesuPrivateTransaction(1);
 
-    final PrivacyTransactionResponse privacyTransactionResponse =
-        privacyController.sendTransaction(transaction);
+    final PrivacySendResponse privacySendResponse = privacyController.sendTransaction(transaction);
 
     final ValidationResult<TransactionInvalidReason> validationResult =
         privacyController.validatePrivateTransaction(
@@ -195,7 +193,7 @@ public class PrivacyControllerTest {
 
     final Transaction markerTransaction =
         privacyController.createPrivacyMarkerTransaction(
-            privacyTransactionResponse.getEnclaveKey(), transaction);
+            privacySendResponse.getEnclaveKey(), transaction);
 
     assertThat(validationResult).isEqualTo(ValidationResult.valid());
     assertThat(markerTransaction.contractAddress()).isEqualTo(PUBLIC_TRANSACTION.contractAddress());
@@ -217,11 +215,10 @@ public class PrivacyControllerTest {
         .thenReturn(ValidationResult.invalid(PRIVATE_NONCE_TOO_LOW));
 
     final PrivateTransaction transaction = buildLegacyPrivateTransaction(0);
-    final PrivacyTransactionResponse privacyTransactionResponse =
-        privacyController.sendTransaction(transaction);
+    final PrivacySendResponse privacySendResponse = privacyController.sendTransaction(transaction);
     final ValidationResult<TransactionInvalidReason> validationResult =
         privacyController.validatePrivateTransaction(
-            transaction, privacyTransactionResponse.getPrivacyGroup());
+            transaction, privacySendResponse.getPrivacyGroup());
     assertThat(validationResult).isEqualTo(ValidationResult.invalid(PRIVATE_NONCE_TOO_LOW));
   }
 
@@ -232,11 +229,10 @@ public class PrivacyControllerTest {
 
     final PrivateTransaction transaction = buildLegacyPrivateTransaction(2);
 
-    final PrivacyTransactionResponse privacyTransactionResponse =
-        privacyController.sendTransaction(transaction);
+    final PrivacySendResponse privacySendResponse = privacyController.sendTransaction(transaction);
     final ValidationResult<TransactionInvalidReason> validationResult =
         privacyController.validatePrivateTransaction(
-            transaction, privacyTransactionResponse.getPrivacyGroup());
+            transaction, privacySendResponse.getPrivacyGroup());
     assertThat(validationResult).isEqualTo(ValidationResult.invalid(INCORRECT_PRIVATE_NONCE));
   }
